TP 3-4 -2

+++ src/main/java/altn72/tpfilrouge/controleur/ProgrammeurService.java
@@ -2,6 +2,8 @@ package altn72.tpfilrouge.controleur;
 
 import altn72.tpfilrouge.modele.Programmeur;
 import altn72.tpfilrouge.modele.ProgrammeurRepository;
+import jakarta.transaction.Transactional;
+import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,41 +11,51 @@ import java.util.Optional;
 
 @Service
 public class ProgrammeurService {
-
-    //@Autowired
     private final ProgrammeurRepository programmeurRepository;
 
     public ProgrammeurService(ProgrammeurRepository programmeurRepository) {
         this.programmeurRepository = programmeurRepository;
     }
 
-    public List<Programmeur> getProgrammeurs(){
+    public List<Programmeur> getProgrammeurs() {
         return programmeurRepository.findAll();
     }
 
-    public Optional<Programmeur> getunProgrammeur(Integer idProgrammeur){
-        return programmeurRepository.findById(idProgrammeur);
+    public Optional<Programmeur> getUnProgrammeur(Integer idProgrammeur) {
+        Optional<Programmeur> unProgrammeur = programmeurRepository.findById(idProgrammeur);
+
+        return Optional.ofNullable(
+                unProgrammeur.orElseThrow(
+                        () -> new IllegalStateException(
+                                "Le programmeur dont l'id est " + idProgrammeur + " n'existe pas")));
     }
 
-    public String deleteProgrammeur(Integer idProgrammeur){
-        try {
+    @Transactional
+    public void supprimerProgrammeur(Integer idProgrammeur) {
+        Optional<Programmeur> unProgrammeur = programmeurRepository.findById(idProgrammeur);
+
+        if (unProgrammeur.isPresent()) {
             programmeurRepository.deleteById(idProgrammeur);
-            return ("Programmeur supprimé avec succès");
-        }
-        catch (Exception e){
-            throw new RuntimeException("Programmeur inexistant");
+        } else {
+            throw new IllegalStateException(
+                    "Le programmeur dont l'id est " + idProgrammeur + " n'existe pas");
         }
     }
 
-    public Programmeur saveProgrammeur(Programmeur programmeur){
-        return programmeurRepository.save(programmeur);
+    @Transactional
+    public void ajouterProgrammeur(Programmeur programmeur) {
+        programmeurRepository.save(programmeur);
     }
 
-    public Programmeur updateProgrammeur(Integer idProgrammeur, Programmeur programmeur){
-        if (programmeurRepository.existsById(idProgrammeur)){
-            deleteProgrammeur(idProgrammeur);
-            return saveProgrammeur(programmeur);
+    @Transactional
+    public void modifierProgrammeur(
+            Integer idProgrammeur,
+            Programmeur programmeurModified) {
+        Programmeur programmeurToModify = programmeurRepository.findById(idProgrammeur).orElseThrow();
+
+        if (programmeurToModify != null) {
+            BeanUtils.copyProperties(programmeurModified, programmeurToModify, "id");
+            programmeurRepository.save(programmeurToModify);
         }
-        throw new RuntimeException("Programmeur inexistant");
     }
 }
